Refactor: 이메일 API 수정
